score/requirements.txt
Django~=3.2.7
mysqlclient~=2.0.1
django-jazzmin~=2.4.8
xlwt~=1.3.0
django-import-export==2.6.1